+++ vm-memory/benches/mmap/mod.rs
@@ -7,11 +7,11 @@
 extern crate criterion;
 extern crate vm_memory;
 
-#[cfg(feature = "rawfd")]
+#[cfg(all(feature = "rawfd", not(target_os = "windows")))]
 use std::fs::{File, OpenOptions};
 use std::mem::size_of;
 
-#[cfg(feature = "rawfd")]
+#[cfg(all(feature = "rawfd", not(target_os = "windows")))]
 use std::path::Path;
 
 use core::hint::black_box;
@@ -88,9 +88,9 @@ pub fn benchmark_for_mmap(c: &mut Criterion) {
     let mut image = make_image(ACCESS_SIZE);
     let buf = &mut [0u8; ACCESS_SIZE];
 
-    #[cfg(feature = "rawfd")]
+    #[cfg(all(feature = "rawfd", not(target_os = "windows")))]
     let mut file = File::open(Path::new("/dev/zero")).expect("Could not open /dev/zero");
-    #[cfg(feature = "rawfd")]
+    #[cfg(all(feature = "rawfd", not(target_os = "windows")))]
     let mut file_to_write = OpenOptions::new()
         .write(true)
         .open("/dev/null")
@@ -116,7 +116,7 @@ pub fn benchmark_for_mmap(c: &mut Criterion) {
             })
         });
 
-        #[cfg(feature = "rawfd")]
+        #[cfg(all(feature = "rawfd", not(target_os = "windows")))]
         c.bench_function(format!("read_from_file_{offset:#0X}").as_str(), |b| {
             b.iter(|| {
                 black_box(&memory)
@@ -165,7 +165,7 @@ pub fn benchmark_for_mmap(c: &mut Criterion) {
             })
         });
 
-        #[cfg(feature = "rawfd")]
+        #[cfg(all(feature = "rawfd", not(target_os = "windows")))]
         c.bench_function(format!("write_to_file_{offset:#0X}").as_str(), |b| {
             b.iter(|| {
                 black_box(&memory)